move hamtwrapper to it's own module
